removed cbayes in favor of breaking up into smaller modules.
